Move Labels component into an own JS module
Rename it from AllLabels to Labels too.
